requirements.txt
torch
tqdm
pillow
transformers
accelerate
bitsandbytes